+++ RELEASENOTES.md
@@ -71,6 +71,9 @@
         buffer bytes.
     *   Add `SkipInfo` to the `AdPlaybackState.AdGroup` to carry skip
         information for each ad in the ad group.
+    *   Fix bug where calling `removeMediaItems(List)` during playing a
+        post-roll created a crash
+        ([#2746](https://github.com/androidx/media/issues/2746)).
 *   Transformer:
     *   Add support for `COMMAND_SET_AUDIO_ATTRIBUTES` and audio focus handling
         in `CompositionPlayer`.